#3 Refactor layout default
