+++ src/pages/index.js
@@ -13,7 +13,15 @@ import {
 } from "../components/Container"
 import List from "../components/List"
 
-const StyledListItem = styled.li`
+const Main = styled.main.attrs(props => ({
+  className: "main",
+}))`
+  grid-column-start: 1;
+  grid-column-end: 2;
+  display: grid;
+`
+
+const StyledExternalLink = styled.a`
   font-weight: 700;
 
   &::after {
@@ -28,65 +36,91 @@ const StyledListItem = styled.li`
   }
 `
 
-const ListItem = props => {
+const ExternalLink = props => {
   return (
-    <StyledListItem>
-      <a href={props.url} rel="me external" title={props.children}>
-        {props.children}
-      </a>
-    </StyledListItem>
+    <StyledExternalLink
+      href={props.url}
+      rel="me external"
+      title={props.children}
+    >
+      {props.children}
+    </StyledExternalLink>
   )
 }
 
 const Home = props => {
   return (
     <Layout>
-      <Container main tertiary role="complementary">
-        <ContainerHeader>
-          <ContainerHeading>
-            Latest articles{" "}
-            <small>
-              in my <Link to="/">blog</Link>
-            </small>
-          </ContainerHeading>
-        </ContainerHeader>
+      <Container as="aside" bright role="banner">
+        <Heading srOnly>Welcome message</Heading>
+        <p>
+          Hi. I’m a <mark>front-end developer</mark> from Sevilla.
+          <br />I write here about things I love: <mark>Javascript</mark>,{" "}
+          <mark>React</mark> and <mark>CSS3</mark>.
+        </p>
 
-        <ContainerMain columnize={3}>
-          <ContainerColumn>This will be a blog post</ContainerColumn>
-          <ContainerColumn>This will be a blog post</ContainerColumn>
-          <ContainerColumn>This will be a blog post</ContainerColumn>
-        </ContainerMain>
+        <p>
+          <mark>Be welcome.</mark>
+        </p>
       </Container>
 
-      <Container main primary>
-        <ContainerHeader>
-          <ContainerHeading>Contact me</ContainerHeading>
-        </ContainerHeader>
-        <ContainerMain columnize={3}>
-          <ContainerColumn>
-            <p>Take a sit. How can I help you?</p>
-          </ContainerColumn>
-          <ContainerColumn>Your name + Email</ContainerColumn>
-          <ContainerColumn>Your message + Send your message</ContainerColumn>
-        </ContainerMain>
-      </Container>
+      <Main>
+        <Container main tertiary role="complementary">
+          <ContainerHeader>
+            <ContainerHeading>
+              Latest articles{" "}
+              <small>
+                in my <Link to="/">blog</Link>
+              </small>
+            </ContainerHeading>
+          </ContainerHeader>
 
-      <Container main bright inline role="contentinfo">
-        <ContainerHeading>I'm also in</ContainerHeading>
-        <List unstyled inlineBlock>
-          <ListItem url="https://github.com/calderon">github</ListItem>
+          <ContainerMain columnize={3}>
+            <ContainerColumn>This will be a blog post</ContainerColumn>
+            <ContainerColumn>This will be a blog post</ContainerColumn>
+            <ContainerColumn>This will be a blog post</ContainerColumn>
+          </ContainerMain>
+        </Container>
 
-          <ListItem url="https://twitter.com/dcalderon">twitter</ListItem>
+        <Container main primary>
+          <ContainerHeader>
+            <ContainerHeading>Contact me</ContainerHeading>
+          </ContainerHeader>
+          <ContainerMain columnize={3}>
+            <ContainerColumn>
+              <p>Take a sit. How can I help you?</p>
+            </ContainerColumn>
+            <ContainerColumn>Your name + Email</ContainerColumn>
+            <ContainerColumn>Your message + Send your message</ContainerColumn>
+          </ContainerMain>
+        </Container>
 
-          <ListItem url="http://linkedin.com/in/danielcalderon">
-            linkedin
-          </ListItem>
-
-          <ListItem url="https://stackoverflow.com/users/7442990/daniel-calder%c3%b3n-portillo">
-            stack overflow
-          </ListItem>
-        </List>
-      </Container>
+        <Container main bright inline role="contentinfo">
+          <ContainerHeading>I'm also in</ContainerHeading>
+          <List unstyled inlineBlock>
+            <li>
+              <ExternalLink url="https://github.com/calderon">
+                github
+              </ExternalLink>
+            </li>
+            <li>
+              <ExternalLink url="https://twitter.com/dcalderon">
+                twitter
+              </ExternalLink>
+            </li>
+            <li>
+              <ExternalLink url="http://linkedin.com/in/danielcalderon">
+                linkedin
+              </ExternalLink>
+            </li>
+            <li>
+              <ExternalLink url="https://stackoverflow.com/users/7442990/daniel-calder%c3%b3n-portillo">
+                stack overflow
+              </ExternalLink>
+            </li>
+          </List>
+        </Container>
+      </Main>
     </Layout>
   )
 }
